+++ CHANGELOG.md
@@ -5,8 +5,8 @@
 - PageLive announces initial info, could be short info & page's snapshot info, after initialized.
 - Modal implementation. Modal will contains almost all info needed related to the active page and PageLive
 - New Shortcut to toogle modal : Ctrl + /
-- TODO: When modal is opened, it will anounce page's snapshot info, continued about how to close the modal
-- TODO: When modal is closed, announce to user.
+- When modal is opened, it will anounce page's snapshot info, continued about how to close the modal
+- When modal is closed, announce to user.
 - TODO: Detect if this is new chat. If not, get the title of current chat.
 - TODO: New shortcut to delete current chat / conversation.
 - TODO: New shortcut to focus on the chat input box.